un poco de Api

+++ libreriaApp/libreriaApp.DAL/Entities/Sales.cs
@@ -10,7 +10,7 @@ namespace libreriaApp.DAL.Entities
         public string stor_id { get; set; }
         public string ord_num { get; set; }
         public DateTime ord_date { get; set; }
-        public int qty { get; set; }
+        public System.Int16 qty { get; set; }
         public string payterms { get; set; }
         public string title_id { get; set; }
